requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.25
aiosqlite==0.19.0
pytz==2024.1
APScheduler==3.10.4
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
click==8.1.7
psycopg2-binary==2.9.9  # для PostgreSQL
python-dateutil==2.8.2
